requirements.txt
# --- Core Backend ---
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==1.10.14
requests==2.32.2
python-jose[bcrypt]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.30
sqlalchemy-utils==0.41.2
PyJWT==2.8.0
anyio==4.3.0
starlette==0.37.2
jinja2==3.1.3
python-dotenv==1.0.1

# --- Machine Learning / AI ---
numpy==1.23.5
pandas==2.0.3
scikit-learn==1.4.1.post1
joblib==1.4.2
imbalanced-learn==0.12.2

xgboost==2.0.3
tensorflow-cpu==2.12.0     # << essential for deployment
torch==2.2.0                # remove or use CPU-only if needed
transformers==4.39.0

# --- Image & PDF Processing ---
pillow==10.3.0
opencv-python==4.8.1.78
matplotlib==3.8.4
scipy==1.10.1
reportlab==4.0.6

# HTML Parsing
beautifulsoup4==4.12.3
